src/requirements.txt
Adafruit-Blinka==8.18.0
boto3==1.26.118
RPLCD==1.3.1
PyYAML==6.0
adafruit_dht
pytest==7.4.0
black==23.3.0